# [pivot] Optimize CLJC pivot exsport for csv files