Optimizations for the latest CLJC reimplementation of pivot exports. Went from these numbers on a sample dataset:
```
6568ms
6158ms
6669ms
```
to:
```
4236ms
4353m
4235ms
```
A flamegraph of the allocation difference: https://flamebin.dev/AGXVBn.